Replaces use of ProcessInformationCopyDictionary with GetProcessInformation.
ProcessInformationCopyDictionary is crashing on 10.7.
This changes WasLaunchedAsLoginOrResumeItem to use lower-level calls
instead of ProcessInformationCopyDictionary, with the hope that it will
bypass the 10.7 crash.  Other reports of crashes in this function on 10.7
claim to have been worked around, so there is reason to believe this will
not crash.

BUG=383029

Review URL: https://codereview.chromium.org/354243003

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 874341f861b3c344324547bee7b9c1b563a52d50