[Mac] Temporarily disable timer coalescing on Mac.
In the interest of conserving power, several incremental changes to timer slack have been committed over the last few weeks.  This CL disables those changes so we can get some numbers for the combined effect in terms of power usage and idle wakeup numbers.

This needs to be checked in because we can only get a good comparative measure for this when run on the bots in a controlled environment.

BUG=356804

Review URL: https://codereview.chromium.org/395763002

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5f41cb5074d900a8332fd72db387e1d48f92f8f3