Bring back the retry limit to 3.
This reverts r284650. There is one test that is flaky 34% of the time, but that is the exception and not the norm. We don't need 12 retries to make browser_tests or other test suites not flaky. If there are very flaky tests, we should disable them instead. I've disabled all the tests that flake more than 5% of the time.

BUG=395189
R=brettw@chromium.org, sergeyberezin@chromium.org

Review URL: https://codereview.chromium.org/413733003

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6986cfc2bf0e211dea91334106c983d25c861810