Take extra lock in base::WaitableEvent::WaitMany.
Currently WaitMany doesn't ensure that the corresponding |Signal| is done
when the WaitMany call is returned.
That is problematic when we share the event over threads and want to wait
for detaching it in order to delete it after that.

BUG=397435
R=agl@chromium.org

Review URL: https://codereview.chromium.org/419773002

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e70af3abb3c92dd43fe668295cc1108f63ffd5a2