Adds some CHECKs to MessagePumpMojo
The thread watcher appears to be kicking on when we're trying to
remove a handle. Not sure why. This converts some DCHECKs to CHECKs
and fixes handling of the deadline. I suspect none of this will help,
but it's worth a shot.

BUG=399769
TEST=none
R=darin@chromium.org

Review URL: https://codereview.chromium.org/454433003

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 44bbf1224d1c0c9a79cef14a2bba4fc06a2cf299