The DisplayVirtualKeyboard function on Windows 8 and beyond should not be displaying the OSK if a physical keyboard is attached to the machine.
Added a function IsKeyboardPresent locally to win_util.cc which uses the setup API's to enumerate
keyboard devices and check for the common ones like ACPI\PNP and HID\VID. This function returns
true if the number of keyboards is greater than 1. This is a hacky approach and is being done
for lack of a good way to determine the presence of a keyboard. The documented methods don't work
well.

BUG=335735

Review URL: https://codereview.chromium.org/644783005

Cr-Original-Commit-Position: refs/heads/master@{#300383}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 74ca738f038557de0e53fd91a8d15774da36f85a