Instrumenting URLRequestJob::NotifyHeadersComplete to locate the source of jankiness
Previous instrumentations showed that URLRequestJob::NotifyHeadersComplete
alone is responsible for ~57.1 janks per hour in IO thread. I need to instrument
the code inside it to find out which part causes jank.

This is a mechanical change that adds instrumentation required to locate the
source of jankiness (i.e. a long-running fragment of code executed as a part of
the task that causes jank) in the code. See the bug for details on what kind of
jank we are after.
A number of similar CLs were landed, and none of them caused issues. They've
helped to find and fix janky code. The code of the instrumentation is highly
optimized and is not expected to affect performance. The code simply creates a
diagnostic task which is identical to ones created by PostTask or IPC message
handlers. The task gets created only in developer build and in Canary channel.

BUG=423948

Review URL: https://codereview.chromium.org/696483002

Cr-Original-Commit-Position: refs/heads/master@{#302694}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 844c155753d7e42619a437ef6f3036aaf466e140