Mechanical rename of base::debug -> base::trace_event [final pass]
This is the last stage of the trace_event directory restructuring.
This is part of a set of 3 CLs which is moving tracing clients to use
the new base::trace_debug namespace.
See crrev.com/837303004 and the related bug for motivations.

BUG=451032
TBR=jam@chromium.org

Review URL: https://codereview.chromium.org/868603007

Cr-Original-Commit-Position: refs/heads/master@{#315397}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 759dd8c5b9b2ece67ac0dc8e406436eff5b69d4f