Delete some tests from allocator_unittests.
The atomics and "die on OOM" bits are tested elsewhere.
The motivation is that this is the only file in all of Chromium that uses
`try` / `catch`.  Originally I wanted to remove just that bit, but when I tried
to compile the file I noticed that the target doesn't exist in most configs,
and the bots don't run the test, so I think the whole thing can go.

Note that no bots run this test. It's only compiled on Windows, and only
in static library builds, and only when building `all` (which means no devs
build it either). http://crbug.com/458309 tracks moving the tests that remain
in this file (which are useful) into base_unittests.

BUG=82385

Review URL: https://codereview.chromium.org/922863002

Cr-Original-Commit-Position: refs/heads/master@{#316084}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a249805f376e40174998be288ed64c6463d69000