MSan: check data is initialized when serializing with base::Pickle.
Serialization implies that the data is going to leave the process one way or
another, and it's never a good idea to leak uninitialized bytes. In most cases
MSan would catch this later on (e.g. on a write()). There are exceptions to
that, however, such as writing to shared memory. It doesn't hurt to have an
extra check here.

BUG=463906
R=thakis@chromium.org

Review URL: https://codereview.chromium.org/988693004

Cr-Original-Commit-Position: refs/heads/master@{#319459}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 2c705b038413d30d7aae5ff5d09882575e93c898