Fix invalid class names passed to ClassLoader.loadClass.
Substitute slashes for dots in the class name if we are going to pass it
to ClassLoader.loadClass via JNI, instead of calling JNIEnv::FindClass.
This avoids spamming the log with warnings from Dalvik about invalid
class names; it was working correctly anyway due to implementation
details of the VM, but we shouldn't be doing this.

This is not the ideal solution to the problem but is a relatively safe
and isolated fix we can cherrypick to a release branch while we work on
a better one.

BUG=461773
R=rmcilroy

Review URL: https://codereview.chromium.org/1008363004

Cr-Original-Commit-Position: refs/heads/master@{#320816}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c16354c567b13a6ce3f9b3d8e4c3af61ee3b8c71