[iOS] WeakNSObject assignment is consistent with copy constructor
The copy constructor was changed (in CL
https://codereview.chromium.org/853503002/) to allow a WeakNSObject to
be copied on a different thread.
This CL is making the assignment operator behave consistently with the
copy constructor: it is now allowed to assign a WeakNSObject on a
different thread.

In any case, dereferencing the WeakNSObject on another thread remains
disallowed.

BUG=394008

Review URL: https://codereview.chromium.org/1014813002

Cr-Original-Commit-Position: refs/heads/master@{#320915}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1fd547a90acc75c995d8ec901ba52879558a27ae